+++ src/conversions.h
@@ -10,9 +10,9 @@ namespace node_tree_sitter {
 void InitConversions();
 v8::Local<v8::Object> RangeToJS(const TSRange &);
 v8::Local<v8::Object> PointToJS(const TSPoint &);
-v8::Local<v8::Number> ByteCountToJS(size_t);
+v8::Local<v8::Number> ByteCountToJS(uint32_t);
 Nan::Maybe<TSPoint> PointFromJS(const v8::Local<v8::Value> &);
-Nan::Maybe<size_t> ByteCountFromJS(const v8::Local<v8::Value> &);
+Nan::Maybe<uint32_t> ByteCountFromJS(const v8::Local<v8::Value> &);
 
 extern Nan::Persistent<v8::String> row_key;
 extern Nan::Persistent<v8::String> column_key;
